Updated libopenmpt to version 0.4.10

+++ Frameworks/OpenMPT/OpenMPT/soundlib/plugins/LFOPlugin.h
@@ -145,7 +145,7 @@ protected:
 
 public:
 	static LFOWaveform ParamToWaveform(float param) { return static_cast<LFOWaveform>(mpt::saturate_round<int>(param * 32.0f)); }
-	static float WaveformToParam(LFOWaveform waveform) { return waveform / 32.0f; }
+	static float WaveformToParam(LFOWaveform waveform) { return static_cast<int>(waveform) / 32.0f; }
 };
 
 OPENMPT_NAMESPACE_END
